Switch first_action -> correct, remove old lines
According to
https://sites.google.com/site/assistmentsdata/home/2012-13-school-data-with-affect,
first_action is whether or not the student took a hint. We actually want
to use the 'correct' column, which encodes hints as incorrect and
actually gives us which questions a student answered correctly. I have
also removed some extra lines that confused the flow of the code.